+++ lib/ocr_scan.dart
@@ -3,6 +3,12 @@ library ocr_scan;
 export 'package:camera/camera.dart';
 export 'package:google_mlkit_commons/google_mlkit_commons.dart';
 export 'package:google_mlkit_text_recognition/google_mlkit_text_recognition.dart';
+export 'package:google_mlkit_barcode_scanning/google_mlkit_barcode_scanning.dart';
 
-export 'src/ocr_scan_preview.dart';
+/// Painter
+export 'src/barcode_scan_zone_painter.dart';
 export 'src/ocr_scan_zone_painter.dart';
+
+/// Main
+export 'src/scan_preview.dart';
+export 'src/zone_painter.dart';